requirements.txt
discord.py>=2.0
python-dotenv>=1.0
aiohttp
yt-dlp>=2024.1.1
ffmpeg-python>=0.2.0